+++ projects/journey-maps-client-testapp/src/app/app.module.ts
@@ -1,4 +1,3 @@
-import {BrowserModule} from '@angular/platform-browser';
 import {NgModule} from '@angular/core';
 
 import {AppComponent} from './app.component';
@@ -15,7 +14,6 @@ import {JourneyMapsClientModule} from '../../../journey-maps-client/src/lib/jour
     SourceHighlightComponent
   ],
   imports: [
-    BrowserModule,
     BrowserAnimationsModule,
     HighlightModule,
     HttpClientModule,