+++ projects/journey-maps-client/src/lib/journey-maps-client.component.spec.ts
@@ -8,6 +8,7 @@ import {MapMarkerService} from './services/map/map-marker.service';
 import {MapInitService} from './services/map/map-init.service';
 import {asyncScheduler, Observable, of, scheduled} from 'rxjs';
 import {JourneyMapsClientModule} from './journey-maps-client.module';
+import {NoopAnimationsModule} from '@angular/platform-browser/animations';
 
 let component: JourneyMapsClientComponent;
 let fixture: ComponentFixture<JourneyMapsClientComponent>;
@@ -102,7 +103,7 @@ describe('JourneyMapsClientComponent#touchEventCollector', () => {
 });
 
 const configureTestingModule = () => TestBed.configureTestingModule({
-  imports: [JourneyMapsClientModule],
+  imports: [JourneyMapsClientModule, NoopAnimationsModule],
   declarations: [JourneyMapsClientComponent],
   providers: [
     Window,