+++ projects/journey-maps-client/src/lib/journey-maps-client.module.ts
@@ -2,9 +2,7 @@ import {NgModule} from '@angular/core';
 import {JourneyMapsClientComponent} from './journey-maps-client.component';
 import {HttpClientModule} from '@angular/common/http';
 import {CommonModule} from '@angular/common';
-import {BrowserModule} from '@angular/platform-browser';
 import {InfoBoxComponent} from './components/info-box/info-box.component';
-import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
 import {SBB_ICON_REGISTRY_PROVIDER, SbbIconModule} from '@sbb-esta/angular-core';
 import {TextInfoBoxComponent} from './components/info-box/components/text-info-box/text-info-box.component';
 import {ButtonInfoBoxComponent} from './components/info-box/components/button-info-box/button-info-box.component';
@@ -28,8 +26,6 @@ import {GlobalStylesComponent} from './components/global-styles/global-styles.co
   ],
   imports: [
     CommonModule,
-    BrowserModule,
-    BrowserAnimationsModule,
     HttpClientModule,
     SbbIconModule,
     SbbButtonModule,
@@ -38,10 +34,7 @@ import {GlobalStylesComponent} from './components/global-styles/global-styles.co
     ReactiveFormsModule,
     AngularResizedEventModule,
   ],
-  providers: [
-    SBB_ICON_REGISTRY_PROVIDER,
-    { provide: Window, useValue: window }
-    ],
+  providers: [SBB_ICON_REGISTRY_PROVIDER],
   exports: [JourneyMapsClientComponent],
 })
 export class JourneyMapsClientModule {